OmniLong-Qwen2.5-VL-7B/demo/requirements.txt
gradio
torch
transformers
pytube
opencv-python
pillow
bitsandbytes
accelerate
torchvision
flash-attn
spaces
num2words